Este contiene cambios en la vista de Register, y muestra función para ocultar y mostrar contraseña
