+++ src/components/login.js
@@ -1,10 +1,5 @@
-import {
-  getAuth,
-  signInWithEmailAndPassword,
-// eslint-disable-next-line import/no-unresolved
-} from 'https://www.gstatic.com/firebasejs/9.6.8/firebase-auth.js';
 // eslint-disable-next-line import/no-cycle
-import { onNavigate } from '../main.js';
+import { signIn } from '../lib/firebase.js';
 
 export const login = `
 <a onclick="onNavigate('/'); return
@@ -27,26 +22,7 @@ function loginFireBase(e) {
   const loginEmail = document.getElementById('loginEmail').value;
   const loginPass = document.getElementById('loginPass').value;
   console.log(loginEmail, loginPass);
-
-  const auth = getAuth();
-  signInWithEmailAndPassword(auth, loginEmail, loginPass)
-    .then((userCredential) => {
-      // Signed in
-      console.log('¡Login Exitoso!');
-      const user = userCredential.user;
-      alert('Logged in!');
-      onNavigate('/feed');
-      // ...
-    })
-    .catch((error) => {
-      const errorCode = error.code;
-      if (errorCode === 'auth/wrong-password') {
-        alert('Tu contraseña es incorrecta, intenta de nuevo o da click en "Olivde mi contraseña"');
-      }
-      if (errorCode === 'auth/invalid-email') {
-        alert('Por favor ingresa un correo válido');
-      }
-    });
+  signIn(loginEmail, loginPass);
 }
 
 window.loginFireBase = loginFireBase;
