+++ src/desktop.css
@@ -19,7 +19,7 @@
         text-align: center;
         }
     .logo{
-        max-width: 100%;
+        max-width: 50%;
     }
     
     .containerButtons{
@@ -80,7 +80,7 @@
         text-decoration: none;
     }
     .containerRegister{
-        margin-top: 6%;
+        margin-top: 10%;
         display:flex;
         flex-direction: column;
         align-items: center;
