+++ src/main.js
@@ -1,9 +1,11 @@
 // Este es el punto de entrada de tu aplicacion
 
-import { home } from './lib/home.js';
-import { login } from './lib/login.js';
-import { register } from './lib/register.js';
-import { feed } from './lib/feed.js'; //
+import { home } from './components/home.js';
+// eslint-disable-next-line import/no-cycle
+import { login } from './components/login.js';
+// eslint-disable-next-line import/no-cycle
+import { register } from './components/register.js';
+import { feed } from './components/feed.js'; //
 // import { getAuth, createUserWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/9.6.8/firebase-auth.js"; //viene desde una CDN y no de lib
 
 const routes = {
